+++ CHANGELOG.md
@@ -7,11 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.5.0] - 2023-01-29
-
-* Dependencies rules: Allow multiple importer packages
+## [0.6.0] - 2023-02-15
 
 ### Added
+
+* "Expensive Loop" rule template and command
+
+## [0.5.0] - 2023-01-29
+
+### Added
+
+* Dependencies rules: Allow multiple importer packages
 
 ## [0.4.0] - 2023-01-05